requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
Pillow>=8.0.0
reportlab>=3.6.0
